pyproject.toml
[build-system]
requires = ["setuptools >= 40.9.0", 'setuptools_scm[toml]>=6.2']
build-backend = "setuptools.build_meta"


[project]
name = "mkdocs_github_changelog"
authors = [{name = "David J Pugh", email="djpugh@gmail.com"}]
maintainers = [{ name = "David J Pugh", email="djpugh@gmail.com" }]
description = "mkdocs extension to autogenerate changelog from github"
readme = {file="README.md", content-type = "text/x-rst"}
keywords = ["mkdocs", "mkdocs-plugin", "github", "changelog", "documentation"]
requires-python = '>3.7'
classifiers = [
        "Environment :: Console",
        "Topic :: Software Development :: Libraries :: Python Modules",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Intended Audience :: Developers",
    	"License :: OSI Approved :: MIT License",
        "Operating System :: Microsoft :: Windows",
        "Operating System :: POSIX :: Linux",
        "Operating System :: MacOS :: MacOS X",
        "Natural Language :: English"
        # "Development Status :: 6 - Mature"
]
dependencies = [
    'ghapi',
    'mkdocs>=1.4',
    "importlib-metadata>=4.6; python_version < '3.10'",
    'typing-extensions; python_version < "3.12"',
    'backports.entry-points-selectable; python_version < "3.10"',
    'importlib_metadata; python_version < "3.11"',
]

dynamic = ["version"]


[project.optional-dependencies]
dev = [
    "setuptools_scm[toml]",
    'importlib_metadata; python_version < "3.8"',
    "nox"
]
dev-lint = [
    "flake8>=5.0.4",
    "flake8-import-order>=0.18.2",
    "flake8-eradicate>=1.4.0",
    "flake8-docstrings>=0.2.7",
    "flake8-pyproject>=1.2.3",
    "flake8-bugbear>=22.0.0",
    "flake8-comprehensions>=3.12.0",
    "flake8-unused-arguments>=0.0.13",
    "flake8-noqa>=1.3.1"
]
dev-security = [
    "pipenv",
    "bandit"
]
dev-types = [
    "mypy",
    "types-colorama",
    "types-setuptools"
]
dev-test= [
    "nskit",
    "pytest>=7.3.1",
    "pytest-cov>=4",
    "pytest-subtests",
    # Run tests in parallel
    "pytest-xdist"
    ]
dev-docs = [
    "mkdocs",
    "mkdocstrings[python]",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-material",
    "mkdocs-git-authors-plugin",
    "mkdocs-coverage",
    "mkdocs-include-markdown-plugin",
    "mike"
]
dev-build = [
    "build"
]

[project.urls]
repository = "https://github.com/djpugh/mkdocs_github_changelog"

[project.scripts]



[project.entry-points."mkdocs.plugins"]
mkdocs_github_changelog = "mkdocs_github_changelog.plugin:MkdocsGithubChangelogPlugin"

[project.entry-points."mkdocs_github_changelog.jinja_environment_factory"]
default = "mkdocs_github_changelog.get_releases:_EnvironmentFactory.default_environment"


# Tools
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = [
    "*.md",
    "*.ini",
    "*.toml",
    "*.whl",
    "*.bicep"
]

[tool.setuptools_scm]
write_to ="src/mkdocs_github_changelog/_version.py"
write_to_template = '__version__ = "{version}"'
version_scheme = "no-guess-dev"

[tool.mypy]
python_version = "3.11"
show_error_codes = true
strict = true
overrides = [
  { module = [
    "virtualenv.*",
    "tox.*",
    "importlib_metadata.*",
    "mkdocs_github_changelog._version"
  ], ignore_missing_imports = true },
]

[tool.flake8]
max-complexity = 22
max-line-length = 120
unused-arguments-ignore-abstract-functions = true
noqa-require-code = true
dictionaries = ["en_US", "python", "technical", "django"]
ignore = [
  "E203", # whitespace before ':'
  "W503", #  line break before binary operator
  "E501",
  "E226",
  "E121",
  "E123",
  "E126",
  "E24",
  "E704"
]
count = true
exclude = [
    # No need to traverse our git directory
    ".git",
    "./.git",
    # There's no value in checking cache directories
    "__pycache__",
    # The conf file is mostly autogenerated, ignore it
    "docs/source/conf.py",
    "./docs/source/conf.py",
    # This contains built code
    "build/*",
    "./build/lib/*",
    "./build/*",
    # This contains the builds that we don't want to check
    "dist",
    # This contains the docs
    "docs/*",
    # Compiled files
    "*.pyc",
    # Docs files
    "*.rst",
    "*.md",
    "*.png",
    ".tox",
    "examples",
    "tests",
    ".venv/*",
    ".tox/*",
    "src/mkdocs_github_changelog/_version.py",
    ".venv*/*"

]
show_source = true
statistics = true
application-import-names="mkdocs_github_changelog"
application-package-names="mkdocs_github_changelog"
docstring-convention="google"
import-order-style="appnexus"

[tool.isort]
profile = "appnexus"
src_paths = ['src/']
known_first_party = [
    "mkdocs_github_changelog.*"
]

known_application = 'mkdocs_github_changelog.*'
force_alphabetical_sort_within_sections = true
force_sort_within_sections = true
reverse_relative = true